Call common/__init__.py from HTTP API
