+++ os/api/__init__.py
@@ -11,7 +11,8 @@ import yaml
 from distutils.dir_util import copy_tree
 
 from flask import Flask, render_template, request, url_for
-from koheron_tcp_client import KClient, command
+from koheron_tcp_client import KClient
+from drivers.common import Common
 
 def log(severity, message):
     print("[" + severity + "] " + message)
@@ -101,50 +102,19 @@ class KoheronAPIApp(Flask):
     def start_client(self):
         self.stop_client()
         self.client = KClient('127.0.0.1', verbose=False)
-        self._init_tcp_server()
+        self.common = Common(self.client)
+        self.common.init()
 
     def stop_client(self):
         if hasattr(self, 'client'):
             self.client.__del__()
 
-    def _init_tcp_server(self):
-        @command('COMMON')
-        def open(self):
-            return self.client.recv_int32()
-
-        open(self)
-
-        @command('COMMON')
-        def init(self): pass
-
-        init(self)
-
-    @command('COMMON')
-    def get_bitstream_id(self):
-        id_array = self.client.recv_buffer(8, data_type='uint32')
-        return ''.join('{:08x}'.format(i) for i in id_array)
-
-    @command('COMMON')
-    def get_dna(self):
-        id_array = self.client.recv_buffer(2, data_type='uint32')
-        return ''.join('{:02x}'.format(i) for i in id_array)
-
-    @command('COMMON', 'I')
-    def set_led(self, value): pass
-
-    @command('COMMON')
-    def get_led(self): 
-        return self.client.recv_uint32()
-
-    @command('COMMON')
-    def ip_on_leds(self): pass
-
     def ping(self):
-        val = self.get_led()
+        val = self.common.get_led()
         for i in range(255):
             time.sleep(0.01)
-            self.set_led(i)
-        self.set_led(val)
+            self.common.set_led(i)
+        self.common.set_led(val)
 
     # ------------------------
     # Instruments
@@ -241,12 +211,12 @@ class KoheronAPIApp(Flask):
 
     def is_bitstream_id_valid(self):
         try:
-            id_ = self.get_bitstream_id()
+            id_ = self.common.get_bitstream_id()
         except:
             log('error', 'Cannot read bitstream ID. Retrying ...')
             try:
                 time.sleep(0.2)
-                id_ = self.get_bitstream_id()
+                id_ = self.common.get_bitstream_id()
             except:
                 log('error', 'Failed to retrieve bitstream ID.')
                 return False
